Conflicts:
	cocos/scripting/auto-generated
	plugin
	samples/Javascript/Shared
	tools/bindings-generator

+++ samples/Cpp/TestCpp/Classes/ShaderTest/ShaderTest2.cpp
@@ -197,8 +197,8 @@ void ShaderSprite::onDraw()
     //
     // Attributes
     //
-#define kQuadSize sizeof(_quad.bl)
-    long offset = (long)&_quad;
+    #define kQuadSize sizeof(_quad.bl)
+    size_t offset = (size_t)&_quad;
     
     // vertex
     int diff = offsetof( V3F_C4B_T2F, vertices);
